feat: move ui to terminal package
