+++ pkg/cmd/apply/apply_test.go
@@ -38,7 +38,7 @@ import (
 	"kusionstack.io/kusion/pkg/engine/runtime"
 	"kusionstack.io/kusion/pkg/engine/runtime/kubernetes"
 	statestorages "kusionstack.io/kusion/pkg/engine/state/storages"
-	"kusionstack.io/kusion/pkg/util/pretty"
+	"kusionstack.io/kusion/pkg/util/terminal"
 	workspacestorages "kusionstack.io/kusion/pkg/workspace/storages"
 )
 
@@ -72,7 +72,7 @@ func NewApplyOptions() *ApplyOptions {
 			NoStyle:      false,
 			Output:       "",
 			IgnoreFields: nil,
-			UI:           pretty.DefaultUI(),
+			UI:           terminal.DefaultUI(),
 		},
 	}
 }
@@ -113,7 +113,7 @@ func mockGenerateSpecWithSpinner() {
 		stack *apiv1.Stack,
 		workspace *apiv1.Workspace,
 		parameters map[string]string,
-		ui *pretty.UI,
+		ui *terminal.UI,
 		noStyle bool,
 	) (*apiv1.Spec, error) {
 		return &apiv1.Spec{Resources: []apiv1.Resource{sa1, sa2, sa3}}, nil
@@ -320,13 +320,13 @@ func mockOperationApply(res models.OpResult) {
 func TestPrompt(t *testing.T) {
 	mockey.PatchConvey("prompt error", t, func() {
 		mockey.Mock((*pterm.InteractiveSelectPrinter).Show).Return("", errors.New("mock error")).Build()
-		_, err := prompt(pretty.DefaultUI())
+		_, err := prompt(terminal.DefaultUI())
 		assert.NotNil(t, err)
 	})
 
 	mockey.PatchConvey("prompt yes", t, func() {
 		mockPromptOutput("yes")
-		_, err := prompt(pretty.DefaultUI())
+		_, err := prompt(terminal.DefaultUI())
 		assert.Nil(t, err)
 	})
 }
